Added Test Cases

+++ updates/apps_add_maintenance_message.php
@@ -8,18 +8,25 @@ class AppsAddMaintenanceMessage extends Migration
 
     public function up()
     {
+        if (Schema::hasColumn('mohsin_mobile_apps', 'maintenance_message')) {
+            return;
+        }
+
         Schema::table('mohsin_mobile_apps', function($table)
         {
-            $table->string('maintenance_message');
+            $table->string('maintenance_message')->default("");
         });
     }
 
     public function down()
     {
-        Schema::table('mohsin_mobile_apps', function($table)
+        if (Schema::hasColumn('mohsin_mobile_apps', 'maintenance_message'))
         {
-            $table->dropColumn('maintenance_message');
-        });
+          Schema::table('mohsin_mobile_apps', function($table)
+          {
+              $table->dropColumn('maintenance_message');
+          });
+        }
     }
 
 }
